% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr_hash.R
\name{learnr_hash}
\alias{learnr_hash}
\title{\strong{learnr} tutorials' add-on elements for hashing}
\description{
The following functions are modified from Colin Rundel's \strong{learnrhash},
available at \url{https://github.com/rundel/learnrhash}, which enable the
encoding, downloading and decoding of hashed answers.

Note that when including these functions in a \strong{learnrhash} Rmd document,
it is necessary that the server function \code{submit_hash_server()} be
included in an R chunk where \code{context="server"} as they interact with
the underlying Shiny functionality. Conversely, the ui function
\code{submit_hash_ui()} must \emph{not} be included in an R chunk with a
\code{context="server"}.
\itemize{
\item \code{submission_server}
\item \code{submission_ui}
}
}
