% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\docType{package}
\name{qfwr}
\alias{qfwr}
\title{qfwr: A package providing some tools for QFwR course.}
\description{
Tools for my course \emph{Quantitative Finance with R} to
facilitate teaching and learning. The interactive exercise
is powered by RStudio's \strong{learnr} and DataCamp's
\strong{testwhat} package.
}
\references{
\itemize{
    \item \href{http://r4ds.had.co.nz/}{R for data science}
    \item \href{https://item.jd.com/12398129.html}{R数据科学}
    \item \href{https://item.jd.com/12069905.html}{量化金融R语言初级教程}
}
}
\author{
Yongyi Zeng <\email{zzyy@xmu.edu.cn}>
}
