% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\name{qfwr_key}
\alias{qfwr_key}
\title{Open an hands-on lab exercise's answer key in code editor.}
\usage{
qfwr_key(lecture = NULL, exercise = NULL)
}
\arguments{
\item{lecture}{An string, such as "L05".}

\item{exercise}{An string, such as "ex1".}
}
\value{
An invisible character vector containing the path of
the answer key file \code{ex#.R}.
}
\description{
Open an hands-on lab exercise's answer key in code editor.
}
\examples{
\dontrun{
qfwr_key("L05", "ex1")
}

}
