% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\name{qfwr_lab}
\alias{qfwr_lab}
\title{Open an hands-on lab's instruction in code editor.}
\usage{
qfwr_lab(lecture = NULL)
}
\arguments{
\item{lecture}{An string, such as "L05".}
}
\value{
An invisible character vector containing the path of the
hands-on lab's instruction RMarkdown file \code{L##-lab.Rmd}.
}
\description{
Open an hands-on lab's instruction in code editor.
}
\examples{
\dontrun{
qfwr_lab("L05")
}

}
