% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\name{qfwr_ls}
\alias{qfwr_ls}
\title{List all interactive exercises in package \strong{qfwr}.}
\usage{
qfwr_ls(pattern = NULL)
}
\arguments{
\item{pattern}{An optional regular expression.}
}
\value{
A character vector containing the names of all
interactive exercises in package \strong{qfwr} with matching
\code{pattern}.
}
\description{
List all interactive exercises which are contained within
R package \strong{qfwr}. \code{qfwr_ls()} is a wrapper of
\code{\link[base]{list.files}}.
}
\examples{
\dontrun{
qfwr_ls(pattern = "^L")
}

}
