% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_praise.R
\name{random_praise}
\alias{random_praise}
\alias{random_praise_cn}
\alias{random_encouragement_cn}
\title{Random praise and encouragement}
\usage{
random_praise_cn(code = FALSE)

random_encouragement_cn(code = FALSE)
}
\arguments{
\item{code}{Random praise for coding exercise (\code{FALSE} by default)}
}
\value{
Character string with a random saying
}
\description{
Random praises and encouragements sayings to
compliment your question and quiz experience.
}
