% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\name{qfwr_ex}
\alias{qfwr_ex}
\title{Run an interactive exercise in package \strong{qfwr}.}
\usage{
qfwr_ex(lecture)
}
\arguments{
\item{lecture}{An string, such as "L05".}
}
\value{
An interactive exercise opened in web browser.
}
\description{
Run an interactive exercise which is contained within R package
\strong{qfwr}. \code{qfwr_ex()} is just a simple wrapper of
\code{learnr::\link[learnr]{run_tutorial}}.
}
\examples{
\dontrun{
qfwr_ex("L05")
}

}
