% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfwr.R
\name{qfwr_ls}
\alias{qfwr_ls}
\title{List all interactive exercises in package \strong{qfwr}.}
\usage{
qfwr_ls()
}
\value{
An invisible data.frame containing "package", "name",
"title", "description", "package_dependencies", "private",
and "yaml_front_matter".
}
\description{
List all interactive exercises which are contained within
R package \strong{qfwr}. \code{qfwr_ls()} is a wrapper of
\code{learnr::\link[learnr]{available_tutorials}}.
}
\examples{
\dontrun{
qfwr_ls()
}

}
